fix_582 Modify test code to add K8S node for configurator
