+++ ohara-configurator/src/test/scala/com/island/ohara/configurator/TestConfiguratorMain.scala
@@ -16,19 +16,12 @@
 
 package com.island.ohara.configurator
 import java.util.concurrent.{Executors, TimeUnit}
-
-import com.island.ohara.agent.fake.FakeK8SClient
-import com.island.ohara.agent.k8s.K8SNodeReport
-import com.island.ohara.client.configurator.v0.NodeApi
 import com.island.ohara.common.rule.OharaTest
 import com.island.ohara.common.util.{CommonUtils, Releasable}
 import com.island.ohara.configurator.Configurator.Mode
 import org.junit.{After, Test}
 import org.scalatest.Matchers._
-
-import scala.concurrent.{Await, ExecutionContext, Future}
-import scala.concurrent.duration._
-import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.{ExecutionContext, Future}
 
 class TestConfiguratorMain extends OharaTest {
 
@@ -64,24 +57,6 @@ class TestConfiguratorMain extends OharaTest {
   def testSshMode(): Unit =
     runMain(Array[String](Configurator.HOSTNAME_KEY, "localhost", Configurator.PORT_KEY, "0"),
             configurator => configurator.mode shouldBe Mode.SSH)
-
-  @Test
-  def testK8SNodes(): Unit = {
-    val k8sClient = new FakeK8SClient(true, None, "") {
-      override def nodes()(implicit executionContext: ExecutionContext): Future[Seq[K8SNodeReport]] =
-        Future.successful(Seq(K8SNodeReport("node1"), K8SNodeReport("node2"), K8SNodeReport("node3")))
-    }
-    val configurator = Configurator.builder.k8sClient(k8sClient).build()
-
-    val nodeApi = NodeApi.access.hostname(configurator.hostname).port(configurator.port)
-    val createNode = Await.result(nodeApi.request.hostname("node1").create(), 15 seconds)
-    createNode.hostname shouldBe "node1"
-
-    val nodes = Await.result(configurator.addK8SNodes(), 15 seconds)
-    nodes.size shouldBe 2
-    nodes(0).hostname shouldBe "node2"
-    nodes(1).hostname shouldBe "node3"
-  }
 
   private[this] def runMain(args: Array[String], action: Configurator => Unit): Unit = {
     Configurator.GLOBAL_CONFIGURATOR_SHOULD_CLOSE = false
